+++ src/App.js
@@ -19,6 +19,7 @@ import ProfilePage from './features/lawyer/profile page/components/ProfilePage';
 import IndividualCasePage from './features/lawyer/individual case page/components/IndividualCasePage';
 import MakeACase from './features/person/make a case/components/MakeACasePage';
 import PersonPage from './features/person/main person page/PersonPage';
+import SendOfferCasePage from './features/lawyer/individual case page/components/SendOfferCasePage';
 
 
 function App() {
@@ -37,6 +38,7 @@ function App() {
             <Route path="notifications" element={<NotificationsList />} />
             <Route path="case-page" element={<IndividualCasePage />} />
             <Route path="profile" element={<ProfilePage />} />
+            <Route path="send-offer-case-page" element={<SendOfferCasePage />} />
           </Route>
           <Route path='/person' element={<PersonPage/>}>
             <Route path="make-a-case" element={<MakeACase />} />
