+++ src/features/lawyer/individual case page/components/SendOfferCasePage.js
@@ -10,13 +10,45 @@ import {
     ListItemIcon,
     ListItemText,
     Button,
+    TextField,
+    FormControl,
+    Input,
+    InputLabel,
+    InputAdornment,
+    OutlinedInput,
+    MenuItem,
+    Select,
+    Chip,
 } from '@mui/material';
+import { useTheme } from '@mui/material/styles';
 import { Link } from 'react-router-dom';
 import LibraryBooksSharpIcon from '@mui/icons-material/LibraryBooksSharp';
 import gandalf from '../../profile page/images/Screenshot 2023-08-04 at 1.28.06.png'
+import { MenuProps, getStyles } from '../style/ChipStyle';
 
 
-function IndividualCasePage() {
+
+function SendOfferCasePage() {
+
+    /* 'Chip' logic:  */
+    const theme = useTheme();
+    const [personName, setPersonName] = React.useState([]);
+
+    const names = [
+        'Rental Contract',
+        'ID Copy',
+        'Authorization',
+        'Copy of passport'
+    ];
+
+    const handleChange = (event) => {
+        const {
+            target: { value },
+        } = event;
+        setPersonName(
+            typeof value === 'string' ? value.split(',') : value,
+        );
+    };
 
     return (
         <>
@@ -88,7 +120,7 @@ function IndividualCasePage() {
                                     Uploaded Documents
                                 </Typography>
                             </Box>
-                            <List sx={{marginLeft: 5}}>
+                            <List sx={{ marginLeft: 5 }}>
                                 <ListItem disablePadding>
                                     <ListItemIcon>
                                         <LibraryBooksSharpIcon />
@@ -112,31 +144,73 @@ function IndividualCasePage() {
                     </Grid>
                 </Grid>
 
-                <Grid container>
+                <Grid container sx={{ marginTop: '20px', marginBottom: '20px', border: '1px solid white', borderRadius: '5px' }}>
                     <Grid item xs={12}>
                         <Box sx={{ width: 1, bgcolor: 'background.paper' }}>
                             <Box sx={{ width: 1, borderBottom: '1px solid white', padding: '10px' }} >
                                 <Typography variant="h6" component="div">
-                                    Uploaded Documents
+                                    Send Offer
                                 </Typography>
                             </Box>
-                            <List sx={{ marginLeft: 5 }}>
-                                
+                            <List sx={{ margin: 2 }}  >
+                                <FormControl fullWidth sx={{ m: 1 }} variant="standard">
+                                    <InputLabel htmlFor="standard-adornment-amount">Offer Amount</InputLabel>
+                                    <Input
+                                        id="standard-adornment-amount"
+                                        startAdornment={<InputAdornment position="start">$</InputAdornment>}
+                                    />
+                                </FormControl>
+                                <TextField
+                                    sx={{ width: '100%', margin: 1 }}
+                                    id="outlined-multiline-static"
+                                    label=""
+                                    multiline
+                                    rows={4}
+                                    defaultValue="Your offer goes here..."
+                                />
+
+                                <FormControl sx={{ m: 1, width: '100%' }}>
+                                    <InputLabel > Upload request: </InputLabel>
+                                    <Select
+                                        labelId="demo-multiple-chip-labe"
+                                        id=''
+                                        multiple
+                                        value={personName}
+                                        onChange={handleChange}
+                                        input={<OutlinedInput id="select-multiple-chip" label="Chip" />}
+                                        renderValue={(selected) => (
+                                            <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.5 }}>
+                                                {selected.map((value) => (
+                                                    <Chip key={value} label={value} />
+                                                ))}
+                                            </Box>
+                                        )}
+                                        MenuProps={MenuProps}
+                                    >
+                                        {names.map((name) => (
+                                            <MenuItem
+                                                key={name}
+                                                value={name}
+                                                style={getStyles(name, personName, theme)}
+                                            >
+                                                {name}
+                                            </MenuItem>
+                                        ))}
+                                    </Select>
+                                </FormControl>
+
+
+                                <Button variant='contained' sx={{ width: 1, margin: 1 }} > Send Offer </Button>
                             </List>
                         </Box>
                     </Grid>
                 </Grid>
 
-                <Grid container>
-                    <>
-                    </>
-                </Grid>
-
                 <Grid container alignItems="center" justifyContent="space-between" sx={{ width: '100%' }} >
-                    <Grid item xs={5} sx={{marginRight: 2}}>
+                    <Grid item xs={5} sx={{ marginRight: 2 }}>
                         <Button variant='contained' sx={{ width: 1 }} > Back </Button>
                     </Grid>
-                    <Grid item xs={5} sx={{marginRight: 0}}>
+                    <Grid item xs={5} sx={{ marginRight: 0 }}>
                         <Button component={Link} to="/lawyer/my-cases/" variant='contained' sx={{ width: 1 }}> Add To My Cases </Button>
                     </Grid>
                 </Grid>
@@ -145,4 +219,4 @@ function IndividualCasePage() {
     )
 };
 
-export default IndividualCasePage;
+export default SendOfferCasePage;
